test: 메서드 분리 및 네이밍 변경
