+++ src/test/java/roomescape/admin/AdminIntegrationTest.java
@@ -1,0 +1,38 @@
+package roomescape.admin;
+
+import static org.hamcrest.CoreMatchers.is;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class AdminIntegrationTest {
+
+    @DisplayName("/admin 요청 시 admin/index.html과 200 응답을 준다.")
+    @Test
+    void when_admin_request() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("/admin/reservation 요청 시 html 정상 응답과 예약 목록을 요청한다.")
+    @Test
+    void when_admin_reservation_request() {
+        RestAssured.given().log().all()
+                .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0));
+    }
+}
